+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * New option `--min-length N%total`
 * New Attribute: `length_m_fraction_longest` what fraction this group is
       compared to the longest group
+* Latest timestamp in the file will be logged & outputted.
 
 ## v3.7.0 (2026-02-08)